% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/McScan.R
\name{bts_cov_stat}
\alias{bts_cov_stat}
\title{Bootstrap of covariance scanning stat for a single change point}
\usage{
bts_cov_stat(X, y, nb = max(dim(X)), st = 0, ed = length(y), bnd = 5)
}
\arguments{
\item{X}{TODO Add Description}

\item{y}{TODO Add Description}

\item{nb}{TODO Add Description}

\item{st}{TODO Add Description}

\item{ed}{TODO Add Description}

\item{bnd}{TODO Add Description}
}
\value{
TODO Add Description
}
\description{
TODO: Description of Details
}
\examples{
# TODO Add Example or remove this
}
