% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence.R
\name{ci_delta}
\alias{ci_delta}
\title{Simultaneous confidence intervals for the differential parameter}
\usage{
ci_delta(
  X,
  y,
  k,
  standardize = FALSE,
  intercept = FALSE,
  delta.hat = NULL,
  nfolds = 3,
  nlambdas = 50,
  alpha = 0.1,
  M = 999,
  do.split = FALSE
)
}
\arguments{
\item{X}{design matrix with the rows containing the observations}

\item{y}{vector of the responses}

\item{k}{location of a change point}

\item{standardize}{boolean; whether to standardise the matrix \code{X}}

\item{intercept}{boolean; whether to include the intercept}

\item{delta.hat}{an estimator of the differential parameter; if \code{delta.hat = NULL}, it is generated via \link[inferchange]{lope}}

\item{nfolds}{number of folds for the cross validation when producing an estimator of the precision matrix of \code{X}}

\item{nlambdas}{size of the grid of the tuning parameter for the cross validation when producing an estimator of the precision matrix of \code{X}}

\item{alpha}{confidence level between 0 and 1}

\item{M}{number of random samples to be generated from the target Gaussian distribution}

\item{do.split}{boolean; whether to split the sample in confidence interval generation or not. When the sample size is moderate, it is recommended not to split the sample.}
}
\value{
an S3 object of class \code{inferchange.ci}, which contains the following fields:
\item{delta.check}{de-sparsified estimator of the differential parameter}
\item{ci}{matrix containing the lower and the upper bound of the simultaneous confidence intervals}
\item{omega}{estimator of the precision matrix of \code{X}}
}
\description{
Generate a de-sparsified estimator of the differential parameter representing the change in the regression coefficients before and after a change point and,
based on a Gaussian approximation result, produces a simultaneous confidence interval at a given level.
}
\details{
See Cho, Kley and Li (2024) for further details.
}
\examples{
\donttest{
}
}
\references{
H. Cho, T. Kley & H. Li (2024) Detection and inference of changes in high-dimensional linear regression with non-sparse structures. \it{arXiv preprint}
}
\seealso{
\link[inferchange]{}
}
