% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLOM.R
\name{clom}
\alias{clom}
\title{Compute CLOM estimator}
\usage{
clom(X, y, k, lambdapath, nfolds = 0, nlambda = 100)
}
\arguments{
\item{X}{covariates; n x p matrix}

\item{y}{response; n vector}

\item{k}{index that splits the n observations into \link{1,k} and (k,n];
{1, ..., n-1}}

\item{lambdapath}{values of lambda to be used when no cross validation is
is done; vector on non-negative numerics}

\item{nfolds}{number of folds to use in cross validation;
if 0 then no cross validation is used; non-negative integer}

\item{nlambda}{number of values on lambdapath for cross validation;
default 100}
}
\value{
TODO Add Description
}
\description{
TODO: Description of Details
}
\examples{
# TODO Add Example or remove this
}
