% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/McScan.R
\name{cp_scan}
\alias{cp_scan}
\title{Covariance Scanning with inf norm}
\usage{
cp_scan(
  X,
  y,
  type = c("bs", "wbs.rand", "not.rand", "wbs.seed", "not.seed"),
  agg = c("inf", "l2", "2s", "2sFWER", "2sFDR"),
  alpha = 0.1,
  nIntv = 10000,
  s = 1,
  ncp = length(y),
  thd = sqrt(log(max(dim(X)))),
  standardise = FALSE,
  stat.bts = NULL,
  bnd = max(round(min(2 * log(length(X)), length(y)/2)), 5),
  post,
  visual = FALSE
)
}
\arguments{
\item{X}{TODO Add Description}

\item{y}{TODO Add Description}

\item{type}{TODO Add Description}

\item{agg}{TODO Add Description}

\item{alpha}{TODO Add Description}

\item{nIntv}{TODO Add Description}

\item{s}{TODO Add Description}

\item{ncp}{TODO Add Description}

\item{thd}{TODO Add Description}

\item{standardise}{TODO Add Description}

\item{stat.bts}{TODO Add Description}

\item{bnd}{TODO Add Description}

\item{post}{TODO Add Description}

\item{visual}{TODO Add Description}
}
\value{
TODO Add Description
}
\description{
Multiscale covariance scanning for data segmentation described in
Section 2.1 of Cho et al. (2024)
}
\examples{
# TODO Add Example or remove this
}
