% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgp.R
\name{dgp_gauss_sparse}
\alias{dgp_gauss_sparse}
\title{Generate a simulated data set}
\usage{
dgp_gauss_sparse(n, p, z, s, rho, sigma = 1)
}
\arguments{
\item{n}{number of observations}

\item{p}{dimension}

\item{z}{length of the first segment}

\item{s}{sparsity}

\item{rho}{L2 norm of delta}

\item{sigma}{standard deviation of errors}
}
\value{
a list with elements X and y
}
\description{
Generate a simulated data set
}
\examples{
data <- dgp_gauss_sparse(n = 20, p = 20, z = 10, s = 3, rho = 1, sigma = 1)
}
