% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/McScan.R
\name{.find_single_cp}
\alias{.find_single_cp}
\title{Find a single change point on (st, ed] with 'bnd' apart from boundaries.}
\usage{
.find_single_cp(CS, st, ed, bnd)
}
\description{
Find a single change point on (st, ed] with 'bnd' apart from boundaries.
}
\keyword{internal}
