% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/McScan.R
\name{find_single_cp}
\alias{find_single_cp}
\title{Find a single change point on given interval}
\usage{
find_single_cp(CS, st, ed, bnd, agg, s, alpha, stat.bts, visual = FALSE)
}
\arguments{
\item{CS}{TODO Add Description}

\item{st}{TODO Add Description}

\item{ed}{TODO Add Description}

\item{bnd}{TODO Add Description}

\item{agg}{TODO Add Description}

\item{s}{TODO Add Description}

\item{alpha}{TODO Add Description}

\item{stat.bts}{TODO Add Description}

\item{visual}{TODO Add Description}
}
\value{
TODO Add Description
}
\description{
Find a single change point on (st, ed] with 'bnd' apart from boundaries.
}
\examples{
# TODO Add Example or remove this
}
