% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferchange-package.R
\docType{package}
\name{inferchange-package}
\alias{inferchange-package}
\title{inferchange: Multiscale Covariance Scanning and Related Algorithms}
\description{
Implements function for inference of changes in high-dimensional linear regression; cf. TODO refer to arxiv preprint.

Implements function for inference of changes in high-dimensional
linear regression; cf. TODO refer to arxiv preprint.
}
\details{
\tabular{ll}{
\cr Package: \tab inferchange
\cr Type:    \tab Package
\cr Version: \tab 0.0.0.9000
\cr Date:    \tab 2024-02-05
\cr License: \tab GPL (>= 3)
}
}
\section{Contents}{

The \pkg{inferchange} package contains functions for the inference
regarding change point analysis.

The main funcions are
\enumerate{
\item \code{McScan},
\item \code{lope},
\item \code{clom}, and
\item \code{ci_delta}.
}
}

\section{Workhorse function}{

A ``workhorse'' function, named `inferchange', that wraps the steps of a full
change point analysis is also available.
}

\references{
Cho, H., Kley, T., and Li, H. (2024).
Detection and inference of changes in high-dimensional linear regression with non-sparse structures
(cf. \url{http://arxiv.org})
}
\seealso{
Useful links:
\itemize{
  \item \url{http://github.com/inferchange/quantspec}
  \item Report bugs at \url{http://github.com/tobiaskley/inferchange/issues}
}

}
\author{
\strong{Maintainer}: Tobias Kley \email{tobias.kley@uni-goettingen.de} (\href{https://orcid.org/0000-0003-4790-9695}{ORCID})

Authors:
\itemize{
  \item Haeran Cho \email{haeran.cho@bristol.ac.uk} (\href{https://orcid.org/0000-0002-0725-8704}{ORCID})
  \item Housen Li \email{housen.li@mathematik.uni-goettingen.de} (\href{https://orcid.org/0000-0002-2434-9878}{ORCID})
}

Other contributors:
\itemize{
  \item Solt Kovacs \email{solt.kovacs@stat.math.ethz.ch} (The function seeded_interval() is a copy of interval() from https://github.com/kovacssolt/ChangePoints) [contributor]
}


Haeran Cho, Tobias Kley, Housen Li
}
\keyword{internal}
