% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferchange.R
\name{inferchange}
\alias{inferchange}
\title{Function for change point detection and inference about changes}
\usage{
inferchange(X, y, ...)
}
\arguments{
\item{X}{Design matrix n x p}

\item{y}{Response vector n}

\item{...}{additional parameters that will be forwarded to \code{[McScan]}}
}
\value{
a named list with elements
\itemize{
\item \code{cp} detected change points, length q
\item \code{delta} n times q matrix, j-th column is \eqn{\\delta_j}
\item \code{ci} n times 2 times q array, \code{ci[,1,j]} and
\code{ci[,1,j]} are lower and upper bounds of the confidence intervals,
respectively
}
}
\description{
This is the \code{workhorse} function that implements all the methodologies
developed in Cho, Kley & Li (2024) that detect multiple change points
and infer about the changes in a high-dimensional linear regression setting.
}
\details{
The following steps are applied:
\itemize{
\item McScan algorithm, cf. Section 2.1 in Cho, Kley & Li (2024),
\item LOPE estimator \eqn{\\hat\\delta_j}, cf. Section 3.1 in Cho, Kley & Li (2024),
\item confidence intervals \eqn{\\C_{ij}(0.1)}, cf. eqn (14),
Section 3.2.2 in Cho, Kley & Li (2024).
}
}
\examples{
old_seed <- .Random.seed
set.seed(12345)
data <- dgp_gauss_sparse(n = 200, p = 20, z = 100, s = 3, rho = 1, sigma = 1)
X <- data$X
y <- data$y
res <- inferchange(X, y)
.Random.seed <- old_seed
}
