% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOPE.R
\name{lope}
\alias{lope}
\title{Compute LOPE estimator}
\usage{
lope(X, y, k, lambdapath, nfolds = 5, nlambda = 100)
}
\arguments{
\item{X}{covariates; n x p matrix}

\item{y}{response; n vector}

\item{k}{index that splits the n observations into {1, ..., k} and
{k+1, ..., n}; k takes values in {1, ..., n-1}}

\item{lambdapath}{values of lambda to be used when no cross validation is
is done; vector of non-negative numerics}

\item{nfolds}{number of folds to use in cross validation;
if 0 then no cross validation is used; non-negative integer}

\item{nlambda}{number of values on lambdapath for cross validation;
default 100}
}
\value{
TODO Add Description
}
\description{
TODO: Description of Details
}
\examples{
# TODO Add Example or remove this
}
