% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/McScan.R
\name{McScan}
\alias{McScan}
\title{Multiscale covariance scanning for data segmentation}
\usage{
McScan(
  X,
  y,
  ncp,
  thd,
  method = c("not", "auto", "wbs", "bs"),
  bnd = max(round(min(2 * log(length(X)), length(y)/2)), 5),
  standardize = FALSE,
  post = TRUE
)
}
\arguments{
\item{X}{Design matrix n x p.}

\item{y}{Response vector n.}

\item{ncp}{Specifies the number of estimated change points, and
overrides the input \code{thd}.}

\item{thd}{Stopping threshold, default log(n*p)/2.}

\item{method}{\code{"not"}  NOT (default) with fixed threshold \code{thd}.
\code{"auto"} NOT with automatic threshold
(overrides \code{ncp} and \code{thd}).
\code{"wbs"}  wild binary segmentation with seeded intervals.
\code{"bs"}   binary segmentation.}

\item{bnd}{At least \code{bnd} away from boundaries of intervals.}

\item{standardize}{Logical, standardizing every coordinate or not.
Default TRUE when \code{thd} is in action.}

\item{post}{Logical, post processing (default) or not.}
}
\value{
A list of class \code{inferchange.cp} with \code{print} and
\code{plot} functions.
}
\description{
The formal definition is given in Section 2.1 of Cho, Kley & Li (2024).
}
\examples{
n = 300
p = 100
s = 10
z = n/3
set.seed(123)
data = dgp_gauss_sparse(n, p, z, s)
est_cps = McScan(data$X, data$y) # default with fixed threshold
plot(est_cps)
}
