% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.inferchange.ci}
\alias{plot.inferchange.ci}
\title{Plotting the confidence intervals for the differential parameters}
\usage{
\method{plot}{inferchange.ci}(x, ...)
}
\arguments{
\item{x}{\code{inferchange.ci} object}

\item{...}{additional arguments}
}
\value{
A plot produced as per the input arguments
}
\description{
Plotting method for S3 objects of class \code{inferchange.ci}.
Displays the de-sparsified estimators of the differential parameters and the simultaneous confidence intervals for the respective parameters
}
\examples{
\donttest{
}
}
\seealso{
\link[inferchange]{ci.delta}
}
