% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/McScan.R
\name{post_process}
\alias{post_process}
\title{Title}
\usage{
post_process(est, cid, CS, agg, alpha, stat.bts, post, visual)
}
\arguments{
\item{est}{TODO Add Description}

\item{cid}{TODO Add Description}

\item{CS}{TODO Add Description}

\item{agg}{TODO Add Description}

\item{alpha}{TODO Add Description}

\item{stat.bts}{TODO Add Description}

\item{post}{TODO Add Description}

\item{visual}{TODO Add Description}
}
\value{
TODO Add Description
}
\description{
Title
}
\examples{
# TODO Add Example or remove this
}
