% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/McScan.R
\name{rnd_interval}
\alias{rnd_interval}
\title{Covariance Scanning}
\usage{
rnd_interval(n, nIntv, minLen = 1)
}
\arguments{
\item{n}{TODO Add Description}

\item{nIntv}{TODO Add Description}

\item{minLen}{TODO Add Description}
}
\value{
TODO Add Description
}
\description{
TODO: Description of Details
}
\examples{
# TODO Add Example or remove this
}
