% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLOM.R
\name{clom}
\alias{clom}
\title{Compute CLOM estimator}
\usage{
clom(
  X,
  y,
  k,
  standardize = FALSE,
  lambdapath = NULL,
  nfolds = 5,
  nlambda = 100
)
}
\arguments{
\item{X}{covariates; n x p matrix}

\item{y}{response; n vector}

\item{k}{index that splits the n observations into {1, ..., k} and
{k+1, ..., n}; k takes values in {1, ..., n-1}}

\item{standardize}{boolean; if \code{standardize = TRUE}, each column of
\code{X} is divided by its L2-norm}

\item{lambdapath}{A user supplied lambda sequence,
only used when no cross validation is done / ignored when nfolds > 1;
vector of non-negative numerics}

\item{nfolds}{number of folds to use in cross validation;
if 0 then no cross validation is used; non-negative integer}

\item{nlambda}{number of values on lambdapath for cross validation;
default 100}
}
\value{
a matrix with p rows and with one (in case of cross-validation) or
length(lambdapath) colummns;
each column corresponds to the estimate obtained for one value of lambda
the values of lambda are available as colnames.
}
\description{
Computes the CLOM estimator \eqn{\hat\delta_{0,n}(k)}
defined in eqn. (8) of Section 3.1 in Cho, Kley & Li (2024).
}
\details{
By default cross-validation is performed and the cross-validation estimate
is returned. for cross-validation the estimates are obtained for all
values of lambda on the lambdapath obtained with \code{\link[=lambdapath]{lambdapath()}}.

If \code{nfolds = 0} is set then the argument \code{lambdapath} has to be
a vector of positive numerics and the estimates for these values of lambda
will be returned. In preprocessing, \code{lambdapath} is sorted to be in
decreasing order.
}
\examples{
# TODO Add Example or remove this
}
