% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lambdapath}
\alias{lambdapath}
\title{Find suitable grid of lambda values}
\usage{
lambdapath(lambda_max, n, p, K = 100)
}
\arguments{
\item{lambda_max}{TODO Add Description}

\item{n}{TODO Add Description}

\item{p}{TODO Add Description}

\item{K}{TODO Add Description}
}
\value{
TODO Add Description
}
\description{
Find suitable grid of lambda values
}
\examples{
# TODO Add Example or remove this
}
