% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/McScan.R
\name{McScan}
\alias{McScan}
\title{Multiscale covariance scanning for data segmentation described in
Section 2.1 of Cho et al. (2024)}
\usage{
McScan(
  X,
  y,
  ncp,
  thd,
  method = c("not", "auto", "wbs", "bs"),
  bnd = max(round(min(2 * log(length(X)), length(y)/2)), 5),
  standardise = FALSE,
  post = TRUE
)
}
\arguments{
\item{X}{Design matrix n x p}

\item{y}{Response vector n}

\item{ncp}{Specifies the number of estimated change points, and
overrides the input \code{thd}}

\item{thd}{Stopping threshold, default log(n*p)/2}

\item{method}{\code{"not"}  NOT (default) with fixed threshold \code{thd}
\code{"auto"} NOT with automatic threshold (overrides \code{ncp} and \code{thd})
\code{"wbs"}  wild binary segmentation with seeded intervals
\code{"bs"}   binary segmentation}

\item{bnd}{At least \code{bnd} away from boundaries of intervals}

\item{standardise}{Logical, standardising every coordinate or not
Default TRUE when \code{thd} is in action}

\item{post}{Logical, post processing (default) or not}
}
\value{
A list of class \code{inferchange.cp} with \code{print} and \code{plot}
}
\description{
Multiscale covariance scanning for data segmentation described in
Section 2.1 of Cho et al. (2024)
}
\examples{
# TODO Add Example or remove this
}
