% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.inferchange.cp}
\alias{plot.inferchange.cp}
<<<<<<< Updated upstream
\title{Plotting the change points}
=======
\title{Plotting the segmentation result returned by \code{McScan}}
>>>>>>> Stashed changes
\usage{
\method{plot}{inferchange.cp}(x, ...)
}
\arguments{
\item{x}{\code{inferchange.cp} object}

\item{...}{additional arguments}
}
\description{
<<<<<<< Updated upstream
Plotting the change points
=======
Plotting the segmentation result returned by \code{McScan}
>>>>>>> Stashed changes
}
