% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.inferchange.cp}
\alias{plot.inferchange.cp}
\title{Plotting the segmentation result}
\usage{
\method{plot}{inferchange.cp}(x, ...)
}
\arguments{
\item{x}{\code{inferchange.cp} object}

\item{...}{additional arguments}
}
\description{
Plotting the segmentation result, which is of class \code{inferchange.cp},
e.g. returned by function \code{McScan}
}
